requirements.txt
azure-functions
azure-appconfiguration
azure-identity
requests
pyyaml
pydantic
python-dateutil
